requirements.txt
streamlit
tensorflow
numpy
Pillow

# streamlit==1.24.1
# tensorflow==2.13.0
# numpy==1.26.4
# Pillow==10.0.0
